% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgovpond-package.R
\docType{package}
\name{bcgovpond-package}
\alias{bcgovpond}
\alias{bcgovpond-package}
\title{bcgovpond: Immutable research data using a data-pond pattern}
\description{
\code{bcgovpond} is an opinionated R package for managing immutable research data
using a data-pond pattern: append-only raw files, explicit metadata, and
stable logical pointers ("views") that decouple analysis code from physical
file names.
}
\details{
Most users will only ever call:
\itemize{
\item \code{\link{create_bcgov_pond_project}} (once, per project)
\item \code{\link{ingest_pond}} (when new data arrives)
\item \code{\link{read_view}} (for analysis)
}

For an overview of the design philosophy and recommended workflow, see the
README on GitHub:
\url{https://github.com/bcgov/bcgovpond}
}
\author{
\strong{Maintainer}: Richard Martin \email{Richard.Martin@gov.bc.ca}

}
\keyword{internal}
