% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.semantic_name}
\alias{.semantic_name}
\title{Extract semantic portion of a filename}
\usage{
.semantic_name(fname)
}
\description{
Removes everything up to and including the first underscore (\verb{_}).
This is used to strip versioning or source prefixes from filenames,
leaving the stable semantic name.
}
