% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{read_view}
\alias{read_view}
\title{Read data backing a view}
\usage{
read_view(semantic_name, view_dir = "data_index/views")
}
\arguments{
\item{semantic_name}{Character scalar. Logical name of the view.}

\item{view_dir}{Directory containing view YAML files.}
}
\value{
A data frame (or tibble).
}
\description{
Resolves the current file backing a view and reads it into R using
an appropriate reader based on file extension.
}
